Add cleanup transform
