+++ modules/460-log-shipper/hooks/testdata/file-to-vector.json
@@ -8,10 +8,18 @@
     }
   },
   "transforms": {
-    "d8_tf_test-vector-dest_00_del_parsed_data": {
+    "d8_tf_test-source_00_clean_up": {
       "drop_on_abort": false,
       "inputs": [
         "test-source"
+      ],
+      "source": "if exists(.pod_labels.\"controller-revision-hash\") {\n    del(.pod_labels.\"controller-revision-hash\")\n}\nif exists(.pod_labels.\"pod-template-hash\") {\n    del(.pod_labels.\"pod-template-hash\")\n}\nif exists(.kubernetes) {\n    del(.kubernetes)\n}\nif exists(.file) {\n    del(.file)\n}",
+      "type": "remap"
+    },
+    "d8_tf_test-vector-dest_00_del_parsed_data": {
+      "drop_on_abort": false,
+      "inputs": [
+        "d8_tf_test-source_00_clean_up"
       ],
       "source": "if exists(.parsed_data) {\n    del(.parsed_data)\n}",
       "type": "remap"
