+++ modules/460-log-shipper/hooks/testdata/namespaced-source.json
@@ -23,7 +23,7 @@
         "process": "process"
       },
       "inputs": [
-        "d8_tf_tests-whispers_whispers-logs_05_log_filter"
+        "d8_tf_tests-whispers_whispers-logs_06_log_filter"
       ],
       "source": "\nfunction process(event, emit)\n\tif event.log.pod_labels == nil then\n\t\temit(event)\n\t\treturn\n\tend\n\tdedot(event.log.pod_labels)\n\temit(event)\nend\nfunction dedot(map)\n\tif map == nil then\n\t\treturn\n\tend\n\tlocal new_map = {}\n\tlocal changed_keys = {}\n\tfor k, v in pairs(map) do\n\t\tlocal dedotted = string.gsub(k, \"%.\", \"_\")\n\t\tif dedotted ~= k then\n\t\t\tnew_map[dedotted] = v\n\t\t\tchanged_keys[k] = true\n\t\tend\n\tend\n\tfor k in pairs(changed_keys) do\n\t\tmap[k] = nil\n\tend\n\tfor k, v in pairs(new_map) do\n\t\tmap[k] = v\n\tend\nend",
       "type": "lua",
@@ -45,45 +45,53 @@
       "source": "if exists(.parsed_data) {\n    del(.parsed_data)\n}",
       "type": "remap"
     },
-    "d8_tf_tests-whispers_whispers-logs_00_log_filter": {
-      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nexists(.parsed_data.foo)",
+    "d8_tf_tests-whispers_whispers-logs_00_clean_up": {
+      "drop_on_abort": false,
       "inputs": [
         "d8_clusterns_source_tests-whispers_tests-whispers_whispers-logs"
       ],
-      "type": "filter"
+      "source": "if exists(.pod_labels.\"controller-revision-hash\") {\n    del(.pod_labels.\"controller-revision-hash\")\n}\nif exists(.pod_labels.\"pod-template-hash\") {\n    del(.pod_labels.\"pod-template-hash\")\n}\nif exists(.kubernetes) {\n    del(.kubernetes)\n}\nif exists(.file) {\n    del(.file)\n}",
+      "type": "remap"
     },
     "d8_tf_tests-whispers_whispers-logs_01_log_filter": {
-      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\n!exists(.parsed_data.foo)",
+      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nexists(.parsed_data.foo)",
       "inputs": [
-        "d8_tf_tests-whispers_whispers-logs_00_log_filter"
+        "d8_tf_tests-whispers_whispers-logs_00_clean_up"
       ],
       "type": "filter"
     },
     "d8_tf_tests-whispers_whispers-logs_02_log_filter": {
-      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nif is_boolean(.parsed_data.foo) || is_float(.parsed_data.foo) {\n    data, err = to_string(.parsed_data.foo);\n    if err != null {\n        false;\n    } else {\n        includes([\"wvrr\"], data);\n    };\n} else {\n    includes([\"wvrr\"], .parsed_data.foo);\n}",
+      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\n!exists(.parsed_data.foo)",
       "inputs": [
         "d8_tf_tests-whispers_whispers-logs_01_log_filter"
       ],
       "type": "filter"
     },
     "d8_tf_tests-whispers_whispers-logs_03_log_filter": {
-      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nif is_boolean(.parsed_data.foo) || is_float(.parsed_data.foo) {\n    data, err = to_string(.parsed_data.foo);\n    if err != null {\n        true;\n    } else {\n        !includes([\"wvrr\"], data);\n    };\n} else {\n    !includes([\"wvrr\"], .parsed_data.foo);\n}",
+      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nif is_boolean(.parsed_data.foo) || is_float(.parsed_data.foo) {\n    data, err = to_string(.parsed_data.foo);\n    if err != null {\n        false;\n    } else {\n        includes([\"wvrr\"], data);\n    };\n} else {\n    includes([\"wvrr\"], .parsed_data.foo);\n}",
       "inputs": [
         "d8_tf_tests-whispers_whispers-logs_02_log_filter"
       ],
       "type": "filter"
     },
     "d8_tf_tests-whispers_whispers-logs_04_log_filter": {
-      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nmatch!(.parsed_data.foo, r'^wvrr$')",
+      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nif is_boolean(.parsed_data.foo) || is_float(.parsed_data.foo) {\n    data, err = to_string(.parsed_data.foo);\n    if err != null {\n        true;\n    } else {\n        !includes([\"wvrr\"], data);\n    };\n} else {\n    !includes([\"wvrr\"], .parsed_data.foo);\n}",
       "inputs": [
         "d8_tf_tests-whispers_whispers-logs_03_log_filter"
       ],
       "type": "filter"
     },
     "d8_tf_tests-whispers_whispers-logs_05_log_filter": {
-      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nif exists(.parsed_data.foo) \u0026\u0026 is_string(.parsed_data.foo) {\n    matched = false\n    matched0, err = match(.parsed_data.foo, r'^wvrr$')\n    if err != null {\n        true\n    }\n    matched = matched || matched0\n    !matched\n} else {\n    true\n}",
+      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nmatch!(.parsed_data.foo, r'^wvrr$')",
       "inputs": [
         "d8_tf_tests-whispers_whispers-logs_04_log_filter"
+      ],
+      "type": "filter"
+    },
+    "d8_tf_tests-whispers_whispers-logs_06_log_filter": {
+      "condition": "if !exists(.parsed_data) {\n    structured, err = parse_json(.message)\n    if err == null {\n        .parsed_data = structured\n    } else {\n        .parsed_data = .message\n    }\n}\n\nif exists(.parsed_data.foo) \u0026\u0026 is_string(.parsed_data.foo) {\n    matched = false\n    matched0, err = match(.parsed_data.foo, r'^wvrr$')\n    if err != null {\n        true\n    }\n    matched = matched || matched0\n    !matched\n} else {\n    true\n}",
+      "inputs": [
+        "d8_tf_tests-whispers_whispers-logs_05_log_filter"
       ],
       "type": "filter"
     }
@@ -92,7 +100,7 @@
     "d8_cluster_sink_loki-storage": {
       "type": "loki",
       "inputs": [
-        "d8_tf_tests-whispers_whispers-logs_05_log_filter"
+        "d8_tf_tests-whispers_whispers-logs_06_log_filter"
       ],
       "healthcheck": {
         "enabled": false
