+++ modules/460-log-shipper/hooks/testdata/throttle.json
@@ -22,7 +22,7 @@
         "process": "process"
       },
       "inputs": [
-        "d8_cluster_source_test-source"
+        "d8_tf_test-source_00_clean_up"
       ],
       "source": "\nfunction process(event, emit)\n\tif event.log.pod_labels == nil then\n\t\temit(event)\n\t\treturn\n\tend\n\tdedot(event.log.pod_labels)\n\temit(event)\nend\nfunction dedot(map)\n\tif map == nil then\n\t\treturn\n\tend\n\tlocal new_map = {}\n\tlocal changed_keys = {}\n\tfor k, v in pairs(map) do\n\t\tlocal dedotted = string.gsub(k, \"%.\", \"_\")\n\t\tif dedotted ~= k then\n\t\t\tnew_map[dedotted] = v\n\t\t\tchanged_keys[k] = true\n\t\tend\n\tend\n\tfor k in pairs(changed_keys) do\n\t\tmap[k] = nil\n\tend\n\tfor k, v in pairs(new_map) do\n\t\tmap[k] = v\n\tend\nend",
       "type": "lua",
@@ -43,6 +43,14 @@
         "d8_tf_test-es-dest_01_ratelimit"
       ],
       "source": "if exists(.parsed_data) {\n    del(.parsed_data)\n}",
+      "type": "remap"
+    },
+    "d8_tf_test-source_00_clean_up": {
+      "drop_on_abort": false,
+      "inputs": [
+        "d8_cluster_source_test-source"
+      ],
+      "source": "if exists(.pod_labels.\"controller-revision-hash\") {\n    del(.pod_labels.\"controller-revision-hash\")\n}\nif exists(.pod_labels.\"pod-template-hash\") {\n    del(.pod_labels.\"pod-template-hash\")\n}\nif exists(.kubernetes) {\n    del(.kubernetes)\n}\nif exists(.file) {\n    del(.file)\n}",
       "type": "remap"
     }
   },
